Adapted Default FloatingPoint DataType to float

+++ 2c.c
@@ -316,7 +316,7 @@ static int signal2scaling_encode(const char *msgname, unsigned id, signal_t *sig
 	assert(copts);
 	const char *type = determine_type(sig->bit_length, sig->is_signed, sig->is_floating);
 	if (sig->scaling != 1.0 || sig->offset != 0.0)
-		type = "double";
+		type = "float";
 	if (copts->use_id_in_name)
 		fprintf(o, "int encode_can_0x%03x_%s(can_obj_%s_t *o, %s in)", id, sig->name, god, copts->use_doubles_for_encoding ? "double" : type);
 	else
@@ -370,7 +370,7 @@ static int signal2scaling_decode(const char *msgname, unsigned id, signal_t *sig
 	assert(copts);
 	const char *type = determine_type(sig->bit_length, sig->is_signed, sig->is_floating);
 	if (sig->scaling != 1.0 || sig->offset != 0.0)
-		type = "double";
+		type = "float";
 	if (copts->use_id_in_name)
 		fprintf(o, "int decode_can_0x%03x_%s(const can_obj_%s_t *o, %s *out)", id, sig->name, god, copts->use_doubles_for_encoding ? "double" : type);
 	else
